Move some parts from Mixins file to utils
